requirements.txt
python-dotenv
google-genai
paho-mqtt>=1.6.1